User registration
we start implementing user registration for our blog application. We are going to put to use our reusable database function to create our user in the database. After creating the user, we'll use the selectOne function to fetch back that user so that we can log them in immediately in the next step