validate posts
we perform validation on the creat posts form. We also add the Publish field to the form enabling the user to check a post as published or not during the creation of the post. When a post is created, we also set a flash message that displays, telling the user that the post was created successfully.